Add files via upload
This is the first upload of the program and all relevant files. It contains the code for the database the prototype of the main GUI as well as a few of the sub GUI including adding new users, adding new shop items, and view all items from a given team. The marketplace code is under development and will search the shop database for matching posts and remove them, a matching pair will be one that has the same name, one is buy one is sell and has a price that is within 20% of each other.