Add files via upload
In this version of the code, the client and server have been implemented along with a GUI for the client to log in, once the user enters the credentials they are sent to the server where they will be compared to the different users in the database. If a user name match is found then the entered password will be checked against the matching user's password.