Add files via upload
The marketplace has been updated so that when a sale is processed the items will be added to sales history with the asset, buyer, seller and price are stored. From the main GUI the see sale history button now displays the content of the sales history so that user can see price history.