Add files via upload
Buying and selling deduct assets from the team's databases, teams cannot list to sell more of an asset than their team has. Posting a sale of an asset will now allow a quantity to be given and will allow the user to choose from a selection of assets based on what is in the asset database.